requirements.txt
pytest==3.3.1
python-telegram-bot==11.1.0
requests==2.20.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
